+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
 Changelog
 =========
 
+- Add Gaussian process classifier for ratio estimation
+
 0.8.4 (2021-06-13)
 ------------------
 - Modify Lotka-Volterra model's priors as many methods do not support discrete random variables.